Se agregó el método `getResultadosActasVolantes` en model Libro
Se introdujo una nueva clase "LibroService" para calcular el desglose de los resultados de "ActaVolante". Modelo `MesaFolio` mejorado y migración para incluir el campo `llamado` y tipos actualizados para claves externas. Se actualizó `ZTestController` para aprovechar estos cambios, asegurando que los datos de `MesaFolio` se creen con precisión.

+++ database/migrations/2024_06_27_002739_create_folio_notas_table.php
@@ -14,7 +14,7 @@ return new class extends Migration {
             $table->integer('escrito')->nullable();
             $table->integer('oral')->nullable();
             $table->integer('definitiva')->nullable();
-            $table->unsignedInteger('operador_id');
+            $table->unsignedBigInteger('operador_id');
             $table->foreignId('acta_volante_id')->nullable();
             $table->foreignUuid('mesa_folio_id');
             $table->foreignId('alumno_id')->nullable();
